fix: use custom-media-element pkg https://github.com/muxinc/custom-media-element
